Sprint6

+++ app/src/main/java/site/pnpl/mira/domain/EmotionCreator.kt
@@ -21,6 +21,8 @@ class EmotionCreator @Inject constructor(
     private val emotionProvider: EmotionProvider
 ) {
 
+    var loadingState: LoadingState<Any> = LoadingState.Loading
+
     private val directoryToSaveEmoji: String
         get() {
             val dir = File("${context.filesDir}$DIRECTORY_FOR_EMOJI")
@@ -33,6 +35,11 @@ class EmotionCreator @Inject constructor(
             val emotionsApi = getEmotionListFromApi()
             val emotionsDb = getEmotionListFromDb().toMutableList()
 
+            if (emotionsApi.isEmpty() && emotionsDb.isEmpty()) {
+                loadingState = LoadingState.Error("")
+                return@launch
+            }
+
             emotionsApi.forEach { emotion ->
                 if (!emotionsDb.remove(emotion)) {
                     saveEmotion(emotion)
@@ -43,6 +50,7 @@ class EmotionCreator @Inject constructor(
                 deleteEmotionsFromBd(emotionsDb)
             }
             emotionProvider.init()
+            loadingState = LoadingState.Success
         }
     }
 
@@ -102,6 +110,14 @@ class EmotionCreator @Inject constructor(
     companion object {
         const val DIRECTORY_FOR_EMOJI = "/emoji/"
     }
+
+
+}
+
+sealed class LoadingState<out T> {
+    object Success : LoadingState<Any>()
+    data class Error(val error: String): LoadingState<Any>()
+    object Loading : LoadingState<Any>()
 }
 
 fun String.parseFileName(): String {
